requirements.txt
alpaca-backtrader-api==0.15.0
finnhub-python==2.4.14
python-dotenv==0.20.0
tensorflow==2.9.1
ratelimit==2.2.1
backoff==2.1.2
